backend/requirements.txt
# APIs y Web
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
httpx==0.27.0

# Base de datos
sqlalchemy==2.0.23
alembic==1.12.1

# Validación y configuración
pydantic[email]==2.4.2
pydantic-settings==2.0.3
python-dotenv==1.0.0

# Procesamiento de medios
pillow==10.1.0
python-magic==0.4.27
exifread==3.0.0
pyheif==0.7.1

# Computer Vision y ML
opencv-python==4.8.1.78
transformers==4.35.2
torch==2.1.1
numpy==1.26.2

# Seguridad
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-asyncio==0.21.1
pytest-mock==3.12.0

# Herramientas de desarrollo
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1
